# webapp/src/main/webapp/docs/swagger.yaml
openapi: 3.0.1
info:
  title: HealthSync API
  version: 1.0.0
...
paths:
...
  /notifications/{notificationId}:
    get:
      operationId: getNotification
      description: Gets the notification
      tags:
        - notifications
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returned notification
          content:
            application/vnd.notification.v1+json:
              schema:
                allOf:
                  - $ref : '#/components/schemas/Notification'
        '403':
          description: Only the user can see its notification
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Notification not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteNotification
      description: Deletes the notification
      tags:
        - notifications
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Ok
        '403':
          description: Only the user can delete its notification
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /notifications:
    get:
      operationId: getNotifications
      description: Gets a page of notifications for the user
      tags:
        - notifications
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            minimum: 0
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Returns page of notifications
          content:
            application/vnd.notifications-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref : '#/components/schemas/Notification'
        '204':
          description: No content for notification page
        '403':
          description: Only the user can see its notifications
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /patients:
    post:
      operationId: createPatient
      description: Creates a patient
      tags:
        - patients
      requestBody:
        content:
          application/vnd.patient.v1+json:
            schema:
              $ref: '#/components/schemas/PatientRegisterForm'
      responses:
        '201':
          description: Patient Created
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already in use
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /patients/{patientId}:
    get:
      operationId: getPatient
      description: Gets the patient
      tags:
        - patients
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returned patient
          content:
            application/vnd.patient.v1+json:
              schema:
                allOf:
                  - $ref : '#/components/schemas/Patient'
        '403':
          description: Only the patient or a doctor that had an appointment with the patient can see the data
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    put:
      operationId: updatePatient
      description: Updates a patient
      tags:
        - patients
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref : '#/components/schemas/PatientEditForm'
      responses:
        '204':
          description: Patient updated
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only the patient can update itself
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Patient not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '409':
          description: Email already in use
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    PatientRegisterForm:
      required:
        - confirmPassword
        - email
        - healthInsurance
        - lastname
        - name
        - password
      type: object
      properties:
        password:
          example: 'johndoesecret123'
          maxLength: 50
          minLength: 4
          pattern: '[a-zA-Z0-9]+'
          type: string
        confirmPassword:
          example: 'johndoesecret123'
          maxLength: 50
          minLength: 4
          pattern: '[a-zA-Z0-9]+'
          type: string
        name:
          example: 'john'
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
        lastname:
          example: 'doe'
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
        email:
          example: 'john@gmail.com'
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z0-9.+-ñÑ]+@[a-zA-Z0-9.-]+(.com|.com.ar|.edu.ar)'
          type: string
        healthInsurance:
          allOf:
            - $ref: '#/components/schemas/HealthInsurance'
          example: 'OSDE'
    PatientEditForm:
      required:
        - email
        - healthInsurance
        - lastname
        - locale
        - name
      type: object
      properties:
        image:
          type: string
          format: binary
        name:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'juan'
        lastname:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'doe'
        email:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z0-9.+-ñÑ]+@[a-zA-Z0-9.-]+(.com|.com.ar|.edu.ar)'
          type: string
          example: 'juan@gmail.com'
        locale:
          type: string
          example: 'es'
        healthInsurance:
          allOf:
            - $ref : '#/components/schemas/HealthInsurance'
...
    Error:
      type: object
      properties:
        message:
          type: string
    ValidationError:
      type: object
      properties:
        attribute:
          type: string
        invalidValue:
          type: string
        message:
          type: string
...
    Patient:
      type: object
      properties:
        appointments:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments?userId=21"
        email:
          type: string
          format: email
          example: "sballerini+10@itba.edu.ar"
        firstName:
          type: string
          example: "Santiago"
        healthInsurance:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/3"
        image:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/images/5"
        lastName:
          type: string
          example: "ballerini"
        notifications:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/notifications?userId=21"
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/patients/21"
    Notification:
      type: object
      properties:
        appointmentId:
          type: integer
          format: int64
          description: The ID of the related appointment.
          example: 400
        id:
          type: integer
          format: int64
          example: 6
        self:
          type: string
          format: uri
          description: The URI for the notification resource.
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/notifications/6"
        userId:
          type: integer
          format: int64
          example: 20
...
    HealthInsurance:
      type: string
      enum:
        - NONE
        - OMINT
        - OSDE
        - SWISS_MEDICAL
        - GALENO
        - MEDIFE
        - MEDICUS
        - ACCORD
        - HOSPITAL_ITALIANO
        - HOSPITAL_ALEMAN
        - HOSPITAL_BRITANICO
        - PREVENCION_SALUD
        - AVALIAN
        - WILLIAM_HOPE
        - PRE_MEDIC
        - OSFE
        - OSPACA
        - ANDAR
        - LUIS_PASTEUR
        - UNION_PERSONAL
        - JERARQUICOS
        - OSDEPYM
        - OSPE
        - OSPLAD
        - OSSEG
        - SANCOR_SALUD
...